+++ src/client/socket.ts
@@ -33,6 +33,8 @@ export interface SocketOptions {
   id?: string
   autoReconnect?: boolean
   reconnectDelayInMs?: number
+  keepAlive?: boolean
+  keepAliveIntervalInMs?: number
   onInvalidMessage?: (message: object) => void
   onReconnect?: () => void
   onError?: (message: object) => void
@@ -46,7 +48,8 @@ const noop = () => {
 type Resolver<T> = (value?: T | PromiseLike<T>) => void
 type Rejecter = (reason?: any) => void
 
-const DEFAULT_RECONNECT_DELAY_IN_MS = 5000
+const DEFAULT_KEEP_ALIVE_INTERVAL_IN_MS = 30000 // 30s
+const DEFAULT_RECONNECT_DELAY_IN_MS = 5000 // 5s
 
 class DefaultEoswsSocket implements EoswsSocket {
   public isConnected: boolean = false
@@ -57,10 +60,7 @@ class DefaultEoswsSocket implements EoswsSocket {
   private listener?: SocketMessageListener
 
   private connectionPromise?: Promise<void>
-  private connectionPromiseResolver?: Resolver<void>
-  private connectionPromiseRejecter?: Rejecter
-
-  private successfulConnectionCount = 0
+  private intervalHandler?: any
 
   private debug: IDebugger
 
@@ -139,6 +139,10 @@ class DefaultEoswsSocket implements EoswsSocket {
     this.isConnected = true
     this.connectionPromise = undefined
 
+    if (this.keepAliveOption() === true) {
+      this.registerKeepAliveHandler()
+    }
+
     this.debug("Signaling completion of `connect` method in the outer scope.")
     resolve()
   }
@@ -146,6 +150,11 @@ class DefaultEoswsSocket implements EoswsSocket {
   private onSocketReconnectOpenFactory = (resolve: Resolver<boolean>) => () => {
     this.debug("Received `onopen` (via reconnect) notification from socket.")
     this.isConnected = true
+    this.connectionPromise = undefined
+
+    if (this.keepAliveOption() === true) {
+      this.registerKeepAliveHandler()
+    }
 
     this.debug("Signaling completion of `reconnect` method in the outer scope.")
     resolve(true)
@@ -157,6 +166,9 @@ class DefaultEoswsSocket implements EoswsSocket {
   private onSocketErrorFactory = (reject: Rejecter) => (event: Event) => {
     this.debug("Received `onerror` notification from socket.")
     this.isConnected = false
+    this.connectionPromise = undefined
+
+    this.cleanSocket()
 
     this.debug("Signaling rejection of `connect` method in the outer scope.")
     reject(event)
@@ -168,10 +180,11 @@ class DefaultEoswsSocket implements EoswsSocket {
   private onSocketClose = (event: CloseEvent) => {
     this.debug("Received `onclose` notification from socket.")
     this.isConnected = false
-    this.cleanSocket()
     this.connectionPromise = undefined
 
-    this.debug("Sending a `onReconnect` notification to client consumer.")
+    this.cleanSocket()
+
+    this.debug("Sending a `onClose` notification to client consumer.")
     this.onClose(event)
 
     if (event.code !== 1000 && event.code !== 1005) {
@@ -212,6 +225,31 @@ class DefaultEoswsSocket implements EoswsSocket {
     return validTypes.indexOf(actualType) > -1
   }
 
+  private registerKeepAliveHandler() {
+    const keepAliveInterval =
+      this.options.keepAliveIntervalInMs || DEFAULT_KEEP_ALIVE_INTERVAL_IN_MS
+
+    this.debug("Unregistering keep alive interval")
+    this.intervalHandler = setInterval(() => {
+      if (!this.isConnected || this.socket === undefined) {
+        return
+      }
+
+      this.debug("Sending keep alive pong through socket.")
+      this.socket.send(JSON.stringify({ type: "pong" }))
+    }, keepAliveInterval)
+  }
+
+  private unregisterKeepAliveHandler() {
+    if (this.intervalHandler === undefined) {
+      return
+    }
+
+    this.debug("Unregistering keep alive interval")
+    clearInterval(this.intervalHandler)
+    this.intervalHandler = undefined
+  }
+
   private async reconnect(): Promise<boolean> {
     if (this.connectionPromise) {
       try {
@@ -243,6 +281,10 @@ class DefaultEoswsSocket implements EoswsSocket {
   }
 
   private cleanSocket() {
+    if (this.intervalHandler !== undefined) {
+      this.unregisterKeepAliveHandler()
+    }
+
     if (this.socket === undefined) {
       return
     }
@@ -252,12 +294,6 @@ class DefaultEoswsSocket implements EoswsSocket {
     this.socket.onerror = noop
     this.socket.onmessage = noop
     this.socket = undefined
-  }
-
-  private cleanConnectionPromise() {
-    this.connectionPromise = undefined
-    this.connectionPromiseRejecter = undefined
-    this.connectionPromiseResolver = undefined
   }
 
   private onInvalidMessage(message: object) {
@@ -274,5 +310,9 @@ class DefaultEoswsSocket implements EoswsSocket {
 
   private onError(message: any) {
     ;(this.options.onError || noop)(message)
+  }
+
+  private keepAliveOption(): boolean {
+    return this.options.keepAlive === undefined ? true : this.options.keepAlive
   }
 }